unit test coverage

+++ domain/model/src/main/java/co/com/bancolombia/model/retotecnicobancolombiaconsumedata/ConsumeData.java
@@ -13,6 +13,6 @@ import lombok.Setter;
 public class ConsumeData {
     String name;
 
-    private double area;
-    private Long population;
+    private long area;
+    private long population;
 }
